+++ app/Models/Module.php
@@ -10,17 +10,43 @@ class Module extends Model
     use HasFactory;
      protected $fillable = ['nome'];
 
+     public function aulas()
+     {
+         return $this->hasMany("App\Models\Aula");
+     }
+
 
      public function verifyAndDelete()
     {
         //realizar alguma validação antes caso seja necessário!!
-      return $this->delete();    
+        $nrAulas = $this->aulas->count();
+        
+        if ($nrAulas > 0):
+            \Session::flash('mensagem', ['type' => 'danger', 'conteudo' => "Aula(s) Relacionada(s) ao Módulo"]);
+            return false;
+        else:
+            return $this->delete();
+        endif;
+         
     }
 
     public static function verifyAndDestroy(array $ids)
     {
         //realizar alguma validação antes caso seja necessário!!
-       return self::destroy($ids);    
+        $nrAulas = \App\Models\Aula::whereIn("module_id", $ids)->count();
+        $nrTotal = $nrAulas +0;
+        $msg = [];
+       
+        if ($nrAulas > 0):
+            $msg[] = "Aula(s) Relacionada(s) ao Módulo";
+        endif;
+        if ($nrTotal > 0):
+            \Session::flash('mensagem', ['type' => 'danger', 'conteudo' => implode("<br>", $msg)]);
+            return false;
+        else:
+            return self::destroy($ids);
+        endif;
+         
     }
 
     public static function getList()
